[TEST] Re-did some requested changes
- I implemented some changes on previous commit, gave me a weird bug
- Could not figure out, so reverted commit and reapplied requested changes
- This is working


Former-commit-id: 2d5fbed6227299f73b0ab90228887bf9c9c0b9ca

+++ tests/client/shm_client.c
@@ -35,8 +35,8 @@ void stop_shm ()
 }
 
 void print_shm ()
-{	
-	print_params((uint32_t) ~0); //prints all devices
+{
+	print_params(~0); //prints all devices
 	print_cmd_map();
 	print_sub_map();
 	print_robot_desc();
